% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle3D.R
\name{angle3D}
\alias{angle3D}
\title{Compute 3D between two points}
\usage{
angle3D(
  x1,
  x2 = NULL,
  x3 = NULL,
  x4 = NULL,
  x5 = NULL,
  x6 = NULL,
  x7 = NULL,
  x8 = NULL,
  x9 = NULL
)
}
\arguments{
\item{x}{Either a single value (the x value of the first point) or a vector of length 9}
}
\value{
Scalar value of Euclidian distance between points
}
\description{
Compute 3D between two points
}
\examples{
a=c(1,2,3)
b=c(3,2,1)
c=c(5,6,7)
angle3D(c(a[1],a[2],a[3],b[1],b[2],b[3],c[1],c[2],c[3]))
}
\references{
https://stackoverflow.com/questions/39671579/compute-euclidean-distance-matrix-from-x-y-z-coordinates
}
