% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle3D.R
\name{fdist}
\alias{fdist}
\title{Example
a=c(1,2,3)
b=c(3,2,1)
c=c(5,6,7)
angle3D(c(a[1],a[2],a[3],
b[1],b[2],b[3],
c[1],c[2],c[3]))}
\usage{
fdist(x1, x2, y1 = NULL, y2 = NULL, z1 = NULL, z2 = NULL)
}
\description{
Example
a=c(1,2,3)
b=c(3,2,1)
c=c(5,6,7)
angle3D(c(a[1],a[2],a[3],
b[1],b[2],b[3],
c[1],c[2],c[3]))
}
