% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmesh.R
\name{plotmesh}
\alias{plotmesh}
\title{Use plotly library to plot a 3d mesh object}
\usage{
plotmesh(mesh, col = NULL, ramp = NULL, title = NULL, ...)
}
\arguments{
\item{x}{a mesh object}
}
\value{
Makes a plotly plot
}
\description{
Use plotly library to plot a 3d mesh object
}
\examples{
  plotmesh(mesh)
plotmesh(mesh,col=mesh$data$hole)
}
\references{
https://plot.ly/r/reference/#mesh3d, https://plot.ly/r/3d-mesh/ https://plot.ly/r/trisurf/ https://laustep.github.io/stlahblog/posts/plotly_trisurf.html
}
