% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox3d.R
\name{st_bbox3d}
\alias{st_bbox3d}
\title{Compute 3D bounding box for a sf object}
\usage{
st_bbox3d(x)
}
\arguments{
\item{x}{an sf object}
}
\value{
a vector representing the 3D bounding box (xmin,ymin,zmin,xmax,ymax,zmax)
}
\description{
Compute 3D bounding box for a sf object
}
