% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2stack.R
\name{df2stack}
\alias{df2stack}
\title{Convert a dataframe to a one-dimensional raster stack with
bands for each column in the dataframe.}
\usage{
df2stack(x)
}
\arguments{
\item{x}{A data.frame object}
}
\value{
A raster stack with a layer for each column in the dataset
}
\description{
Convert a dataframe to a one-dimensional raster stack with
bands for each column in the dataframe.
}
\examples{
n=5
# Create a test data.frame to illustrate the function
x=data.frame(x=1:n,
            y=rep(1:2,each=2,len=n),
            pres=rbinom(n,size=1,p=0.1),
            v1=rnorm(n),
            v2=rnorm(n),
            v3=rnorm(n))
# Convert to a raster stack
x2=df2stack(x)
# show the coordinates of the raster
coordinates(x2)
plot(x2)
}
